Adding singleProduct reducer to combineReducer + moving SingleProduct to product folder
