+++ client/components/index.js
@@ -5,5 +5,5 @@
  */
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
-export {default as SingleProduct} from './SingleProduct'
+export {default as SingleProduct} from './Products/SingleProduct'
 export {Login, Signup} from './auth-form'
