+++ client/store/index.js
@@ -4,8 +4,13 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import productsReducer from './products'
+import singleProduct from './singleProduct'
 
-const reducer = combineReducers({user, products: productsReducer})
+const reducer = combineReducers({
+  user,
+  products: productsReducer,
+  singleProduct: singleProduct
+})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
