+++ client/store/singleProduct.js
@@ -11,9 +11,19 @@ const gotSingleProduct = singleProduct => {
   }
 }
 
+const imageUrl =
+  'https://upload.wikimedia.org/wikipedia/commons/thumb/1/1b/Creative-Tail-Animal-dog.svg/256px-Creative-Tail-Animal-dog.svg.png'
+
 // THUNKS
 export const fetchOneProduct = id => async dispatch => {
-  const {data} = await axios.get(`/api/products/${id}`)
+  // const {data} = await axios.get(`/api/products/${id}`)
+  const data = {
+    id: 1,
+    name: 'Dragon',
+    description: 'get three dragons for the price of one!',
+    imageUrl: imageUrl,
+    price: 1000
+  }
   dispatch(gotSingleProduct(data))
 }
