+++ server/api/products.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const Product = require('../db/product')
+const Product = require('../db/models/product')
 
 router.get('/', async (req, res, next) => {
   try {
