+++ server/db/models/index.js
@@ -1,46 +1,46 @@
-// const db = require('./database')
+// const db = require('../db')
 
-// // const Order = require('./order')
-// // const OrderList = require('./orderList')
-// // const Cart = require('./cart')
+// const Order = require('./order')
+// const OrderList = require('./orderList')
+// const Cart = require('./cart')
 
-// // const Session = require('./session')
-// // const User = require('./user')
-// // const Review = require('./review')
-// // const Product = require('./product')
+// const Session = require('./session')
+// const User = require('./user')
+// const Review = require('./review')
+// const Product = require('./product')
 
 // // Model Associations
-// // Cart.hasMany(OrderList)
-// // OrderList.belongsTo(Cart)
+// Cart.hasMany(OrderList)
+// OrderList.belongsTo(Cart)
 
-// // Cart.hasOne(Session)
-// // Session.belongsTo(Cart)
+// Cart.hasOne(Session)
+// Session.belongsTo(Cart)
 
-// // User.hasMany(Review)
-// // Review.belongsTo(User)
+// User.hasMany(Review)
+// Review.belongsTo(User)
 
-// // User.hasOne(Session)
-// // Session.belongsTo(User)
+// User.hasOne(Session)
+// Session.belongsTo(User)
 
-// // User.hasMany(Order)
-// // Order.belongsTo(User)
+// User.hasMany(Order)
+// Order.belongsTo(User)
 
-// // Product.hasMany(Review)
-// // Review.belongsTo(Product)
+// Product.hasMany(Review)
+// Review.belongsTo(Product)
 
-// // OrderList.hasMany(Product)
-// // Product.belongsTo(OrderList)
+// OrderList.hasMany(Product)
+// Product.belongsTo(OrderList)
 
-// // Order.hasMany(OrderList)
-// // OrderList.belongsTo(Order)
+// Order.hasMany(OrderList)
+// OrderList.belongsTo(Order)
 
 // module.exports = {
 //   db,
-//   // Order,
-//   // OrderList,
-//   // Cart,
-//   // Session,
-//   // User,
-//   // Review,
-//   // Product
+//   Order,
+//   OrderList,
+//   Cart,
+//   Session,
+//   User,
+//   Review,
+//   Product
 // }
